+++ frontend/src/UpdateQuestionForm/UpdateQuestionForm.tsx
@@ -6,12 +6,12 @@ import Alert from "@mui/material/Alert";
 import AlertTitle from "@mui/material/AlertTitle";
 import CircularProgress from "@mui/material/CircularProgress";
 
-import { FormInputText } from "../form_components/FormInputText";
-import { FormInputDropdown } from "../form_components/FormInputDropdown";
-import { FormMultipleInputDropdown } from "../form_components/FormMultipleInputDropdown";
-import FormInputTextEditor from "../form_components/FormInputTextEditor";
+import { FormInputText } from "../components/form_components/FormInputText";
+import { FormInputDropdown } from "../components/form_components/FormInputDropdown";
+import { FormMultipleInputDropdown } from "../components/form_components/FormMultipleInputDropdown";
+import FormInputTextEditor from "../components/form_components/FormInputTextEditor";
 
-import { getCurrentUser } from "../services/auth.service";
+import { getCurrentUser } from "../utils/auth.service";
 
 interface Question {
   _id: string;
