+++ frontend/src/components/AdminProtected/AdminProtected.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from 'react';
 import useAdminCheck from './useAdminCheck';
-import ErrorPage from '../ErrorPage';
+import ErrorPage from '../../pages/ErrorPage';
 import CircularProgress from "@mui/material/CircularProgress";
 
 interface AdminProtectedProps {
