--- frontend/src/form_components/FormStyles.css
@@ -1,8 +1,0 @@
-/* .root {
-padding: 50px;
-height: 100vh;
-}
-
-.paper {
-background-color: 'white';
-} */
