+++ frontend/src/pages/Matching/CodeSpace.tsx
@@ -5,8 +5,8 @@ import CodeMirror from '@uiw/react-codemirror';
 import { javascript } from '@codemirror/lang-javascript';
 import { langs } from '@uiw/codemirror-extensions-langs';
 import ScrollToBottom from 'react-scroll-to-bottom';
-import { addHistory } from "../../services/user-history.service";
-import { runcode } from "../../services/code.service";
+import { addHistory } from "../../utils/user-history.service";
+import { runcode } from "../../utils/code.service";
 import logo from '../../images/peerPrepLogo.png';
 import PublishIcon from '@mui/icons-material/Publish';
 import LogoutIcon from '@mui/icons-material/Logout';
